feat: add option to merge same classes that are after each other
The merge_blocks() method has been added to the Schedule class to merge all same class blocks that are after each other and remove breaks between them. This allows for a more condensed and organized representation of the schedule.

The merge option can be enabled by passing the "-m" or "--merge" flag when running the script. This will merge the same classes that are after each other in the schedule.

Example usage:
python script.py -n -i -m /path/to/file.html